This is a snapshot from http://www.suckless.org/hg.rc/dwm on Mar 18, 2008.
This is the base for the changes that I will be making. A huge thanks to all
those who have contributed to dwm. I have really enjoyed learning more about
how window managers work in X, and thanks to the nice small codebase in dwm
I have found it relatively easy to make the changes I need to create a
window manager that does what I need it to.

Thanks dwm.